[Silver II] Title: 새끼치기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/17291. 새끼치기/새끼치기.cc
@@ -7,7 +7,7 @@ int main() {
 	cin.tie(nullptr);
 	int n = 0;
 	cin >> n;
-	vector<long long> arr(21,0);
+	vector<int> arr(21,0);
 	arr[1] = 1;
 	arr[2] = 2;
 	arr[3] = 4;
